+++ src/constants/actionTypes.js
@@ -1,15 +1,15 @@
-export const RESET_GAME = 'RESET_GAME';
+export const RESET_GAME = 'common/RESET_GAME';
 
-export const TOGGLE_ALL_CARDS = 'TOGGLE_ALL_CARDS';
-export const OPEN_CARD = 'OPEN_CARD';
-export const DISABLE_CARD = 'DISABLE_CARD';
-export const DISACTIVE_ALL_CARDS = 'DISACTIVE_ALL_CARDS';
+export const TOGGLE_ALL_CARDS = 'cards/TOGGLE_ALL_CARDS';
+export const OPEN_CARD = 'cards/OPEN_CARD';
+export const DISABLE_CARD = 'cards/DISABLE_CARD';
+export const DISACTIVE_ALL_CARDS = 'cards/DISACTIVE_ALL_CARDS';
 
-export const SET_GAME_ACTION = 'SET_GAME_ACTION';
-export const SET_COUNTER = 'SET_COUNTER';
-export const SET_GAME_RESULT = 'SET_GAME_RESULT';
-export const SET_GAME_STATE = 'SET_GAME_STATE';
-export const PICTURES_FETCHING = 'PICTURES_FETCHING';
+export const SET_GAME_ACTION = 'game_state/SET_GAME_ACTION';
+export const SET_COUNTER = 'game_state/SET_COUNTER';
+export const SET_GAME_RESULT = 'game_state/SET_GAME_RESULT';
+export const SET_GAME_STATE = 'game_state/SET_GAME_STATE';
+export const PICTURES_FETCHING = 'game_state/PICTURES_FETCHING';
 
-export const SET_GAME_SETTINGS = 'SET_GAME_SETTINGS';
-export const SET_CARDS = 'SET_CARDS';
+export const SET_GAME_SETTINGS = 'game_settings/SET_GAME_SETTINGS';
+export const SET_CARDS = 'game_settings/SET_CARDS';
