+++ src/index.pug
@@ -5,10 +5,10 @@ html(lang="en")
 		title MEMORY STORM 2
 	body
 		#root
-		#splash
-			.title MEMORY STORM
-				.exclamation work hard , play hard
-					div !
-			.splash-wrap
-				- for(var i = 0; i < 100; i++)
-					.splash
+		//- #splash
+		//- 	.title MEMORY STORM
+		//- 		.exclamation work hard , play hard
+		//- 			div !
+		//- 	.splash-wrap
+		//- 		- for(var i = 0; i < 100; i++)
+		//- 			.splash
